flytt til egen klient
